* Added the logic for the case where drivers data and vehicle consumption data is not static. * Cleaned up the code by segregation of functions in utils and create_tables file.
